Train model first before being able to use model
