+++ src/router/Router.svelte
@@ -14,7 +14,7 @@
   import ModelPage from '../pages/model/ModelPage.svelte';
   import PlaygroundPage from '../pages/PlaygroundPage.svelte';
   import TrainingPage from '../pages/training/TrainingPage.svelte';
-  import OutputPage from '../pages/OutputPage.svelte';
+  import OutputPage from '../pages/output/OutputPage.svelte';
   import { currentPageComponent } from '../views/currentComponentStore';
   import { currentPath, isValidPath, navigate, Paths, PathType } from './paths';
